+++ AllReadyApp/Web-App/AllReady/Security/AllReadyOpenCorsPolicy.cs
@@ -1,0 +1,24 @@
+﻿using Microsoft.AspNetCore.Cors.Infrastructure;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace AllReady.Security
+{
+    public static class AllReadyCorsPolicyFactory
+    {
+        public static CorsPolicy BuildAllReadyOpenCorsPolicy()
+        {
+            var corsBuilder = new CorsPolicyBuilder();
+
+            corsBuilder.AllowAnyHeader();
+            corsBuilder.AllowAnyMethod();
+            // we need to revisit this and build a whitelise of IP addresses
+            corsBuilder.AllowAnyOrigin();
+            corsBuilder.AllowCredentials();
+
+            return corsBuilder.Build();
+        }
+    }
+}
