+++ AllReadyApp/Web-App/AllReady/Startup.cs
@@ -1,23 +1,10 @@
 using System;
-using System.Collections.Generic;
 using System.Globalization;
-using System.Reflection;
-using System.Threading.Tasks;
-using AllReady.Areas.Admin.ViewModels.Validators;
-using AllReady.Areas.Admin.ViewModels.Validators.Task;
-using AllReady.Controllers;
 using AllReady.DataAccess;
 using AllReady.Hangfire;
 using AllReady.Models;
-using AllReady.Providers;
-using AllReady.Providers.ExternalUserInformationProviders;
-using AllReady.Providers.ExternalUserInformationProviders.Providers;
 using AllReady.Security;
-using AllReady.Services;
 using Autofac;
-using Autofac.Extensions.DependencyInjection;
-using Autofac.Features.Variance;
-using MediatR;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
@@ -27,22 +14,11 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.PlatformAbstractions;
-using AllReady.Security.Middleware;
 using Newtonsoft.Json.Serialization;
-using Microsoft.AspNetCore.Cors.Infrastructure;
 using Hangfire;
-using Hangfire.SqlServer;
 using AllReady.ModelBinding;
-using AllReady.Services.Mapping;
-using AllReady.Services.Mapping.GeoCoding;
-using AllReady.Services.Mapping.Routing;
 using Microsoft.AspNetCore.Localization;
-using AllReady.Services.Twitter;
-using CsvHelper;
-using AllReady.Services.Sms;
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authentication.OAuth;
-using Microsoft.AspNetCore.Authentication.Twitter;
+using AllReady.Configuration;
 
 namespace AllReady
 {
@@ -86,14 +62,9 @@ namespace AllReady
         {
             //Add CORS support.
             // Must be first to avoid OPTIONS issues when calling from Angular/Browser
-            var corsBuilder = new CorsPolicyBuilder();
-            corsBuilder.AllowAnyHeader();
-            corsBuilder.AllowAnyMethod();
-            corsBuilder.AllowAnyOrigin();
-            corsBuilder.AllowCredentials();
             services.AddCors(options =>
             {
-                options.AddPolicy("allReady", corsBuilder.Build());
+                options.AddPolicy("allReady", AllReadyCorsPolicyFactory.BuildAllReadyOpenCorsPolicy());
             });
 
             // Add Application Insights data collection services to the services container.
@@ -102,17 +73,7 @@ namespace AllReady
             // Add Entity Framework services to the services container.
             services.AddDbContext<AllReadyContext>(options => options.UseSqlServer(Configuration["Data:DefaultConnection:ConnectionString"]));
 
-            services.Configure<AzureStorageSettings>(Configuration.GetSection("Data:Storage"));
-            services.Configure<DatabaseSettings>(Configuration.GetSection("Data:DefaultConnection"));
-            services.Configure<EmailSettings>(Configuration.GetSection("Email"));
-            services.Configure<SampleDataSettings>(Configuration.GetSection("SampleData"));
-            services.Configure<GeneralSettings>(Configuration.GetSection("General"));
-            services.Configure<GetASmokeAlarmApiSettings>(Configuration.GetSection("GetASmokeAlarmApiSettings"));
-            services.Configure<TwitterAuthenticationSettings>(Configuration.GetSection("Authentication:Twitter"));
-            services.Configure<TwilioSettings>(Configuration.GetSection("Authentication:Twilio"));
-            services.Configure<MappingSettings>(Configuration.GetSection("Mapping"));
-            services.Configure<ApprovedRegionsSettings>(settings => settings.Enabled = false);
-            services.Configure<ApprovedRegionsSettings>(Configuration.GetSection("ApprovedRegions"));
+            Options.LoadConfigurationOptions(services, Configuration);
 
             // Add Identity services to the services container.
             services.AddIdentity<ApplicationUser, IdentityRole>(options =>
@@ -158,103 +119,10 @@ namespace AllReady
             services.AddHangfire(configuration => configuration.UseSqlServerStorage(Configuration["Data:HangfireConnection:ConnectionString"]));
 
             // configure IoC support
-            var container = CreateIoCContainer(services);
+            var container = AllReady.Configuration.Services.CreateIoCContainer(services, Configuration);
             return container.Resolve<IServiceProvider>();
         }
 
-        private IContainer CreateIoCContainer(IServiceCollection services)
-        {
-            // todo: move these to a proper autofac module
-            // Register application services.
-            services.AddSingleton(x => Configuration);
-            services.AddTransient<IEmailSender, AuthMessageSender>();
-            services.AddTransient<ISmsSender, AuthMessageSender>();
-            services.AddTransient<IDetermineIfATaskIsEditable, DetermineIfATaskIsEditable>();
-            services.AddTransient<IValidateEventEditViewModels, EventEditViewModelValidator>();
-            services.AddTransient<ITaskEditViewModelValidator, TaskEditViewModelValidator>();
-            services.AddTransient<IItineraryEditModelValidator, ItineraryEditModelValidator>();
-            services.AddTransient<IOrganizationEditModelValidator, OrganizationEditModelValidator>();
-            services.AddTransient<IRedirectAccountControllerRequests, RedirectAccountControllerRequests>();
-            services.AddSingleton<IImageService, ImageService>();
-            services.AddSingleton<ICsvFactory, CsvFactory>();
-            services.AddTransient<SampleDataGenerator>();
-            services.AddSingleton<IHttpClient, StaticHttpClient>();
-            services.AddSingleton<ITwitterService, TwitterService>();
-
-            if (Configuration["Mapping:EnableGoogleGeocodingService"] == "true")
-            {
-                services.AddSingleton<IGeocodeService,GoogleGeocodeService>();
-            }
-            else
-            {
-                services.AddSingleton<IGeocodeService, FakeGeocodeService>();
-            }
-
-            if (Configuration["Data:Storage:EnableAzureQueueService"] == "true")
-            {
-                // This setting is false by default. To enable queue processing you will 
-                // need to override the setting in your user secrets or env vars.
-                services.AddTransient<IQueueStorageService, QueueStorageService>();
-            }
-            else
-            {
-                // this writer service will just write to the default logger
-                services.AddTransient<IQueueStorageService, FakeQueueWriterService>();
-            }
-
-            if (Configuration["Authentication:Twilio:EnableTwilio"] == "true")
-            {
-                services.AddSingleton<IPhoneNumberLookupService, TwilioPhoneNumberLookupService>();
-                services.AddSingleton<ITwilioWrapper, TwilioWrapper>();
-            }
-            else
-            {
-                services.AddSingleton<IPhoneNumberLookupService, FakePhoneNumberLookupService>();
-            }
-            
-            var containerBuilder = new ContainerBuilder();
-            containerBuilder.RegisterSource(new ContravariantRegistrationSource());
-            containerBuilder.RegisterAssemblyTypes(typeof(IMediator).GetTypeInfo().Assembly).AsImplementedInterfaces();
-            containerBuilder.RegisterAssemblyTypes(typeof(Startup).GetTypeInfo().Assembly).AsImplementedInterfaces();
-            containerBuilder.Register<SingleInstanceFactory>(ctx =>
-            {
-                var c = ctx.Resolve<IComponentContext>();
-                return t => c.Resolve(t);
-            });
-
-            containerBuilder.Register<MultiInstanceFactory>(ctx =>
-            {
-                var c = ctx.Resolve<IComponentContext>();
-                return t => (IEnumerable<object>)c.Resolve(typeof(IEnumerable<>).MakeGenericType(t));
-            });
-
-            //ExternalUserInformationProviderFactory registration
-            containerBuilder.RegisterType<TwitterExternalUserInformationProvider>().Named<IProvideExternalUserInformation>("Twitter");
-            containerBuilder.RegisterType<GoogleExternalUserInformationProvider>().Named<IProvideExternalUserInformation>("Google");
-            containerBuilder.RegisterType<MicrosoftAndFacebookExternalUserInformationProvider>().Named<IProvideExternalUserInformation>("Microsoft");
-            containerBuilder.RegisterType<MicrosoftAndFacebookExternalUserInformationProvider>().Named<IProvideExternalUserInformation>("Facebook");
-            containerBuilder.RegisterType<ExternalUserInformationProviderFactory>().As<IExternalUserInformationProviderFactory>();
-
-            //Hangfire
-            containerBuilder.Register(icomponentcontext => new BackgroundJobClient(new SqlServerStorage(Configuration["Data:HangfireConnection:ConnectionString"])))
-                .As<IBackgroundJobClient>();
-
-            //auto-register Hangfire jobs by convention
-            //http://docs.autofac.org/en/latest/register/scanning.html
-            var assembly = typeof(Startup).GetTypeInfo().Assembly;
-            containerBuilder
-                .RegisterAssemblyTypes(assembly)
-                .Where(t => t.Namespace == "AllReady.Hangfire.Jobs" && t.GetTypeInfo().IsInterface)
-                .AsImplementedInterfaces();
-
-            containerBuilder.RegisterType<GoogleOptimizeRouteService>().As<IOptimizeRouteService>().SingleInstance();
-
-            //Populate the container with services that were previously registered
-            containerBuilder.Populate(services);
-
-            var container = containerBuilder.Build();
-            return container;
-        }
 
         // Configure is called after ConfigureServices is called.
         public async void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory, SampleDataGenerator sampleData, AllReadyContext context, IConfiguration configuration)
@@ -315,83 +183,7 @@ namespace AllReady
 
             app.UseRequestLocalization();
 
-            // Add cookie-based authentication to the request pipeline.
-            app.UseIdentity();
-
-            // Add token-based protection to the request inject pipeline
-            app.UseTokenProtection(new TokenProtectedResourceOptions
-            {
-                Path = "/api/request",
-                PolicyName = "api-request-injest"
-            });
-
-            // Add authentication middleware to the request pipeline. You can configure options such as Id and Secret in the ConfigureServices method.
-            // For more information see http://go.microsoft.com/fwlink/?LinkID=532715
-            if (Configuration["Authentication:Facebook:AppId"] != null)
-            {
-                var options = new FacebookOptions
-                {
-                    AppId = Configuration["Authentication:Facebook:AppId"],
-                    AppSecret = Configuration["Authentication:Facebook:AppSecret"],
-                    BackchannelHttpHandler = new FacebookBackChannelHandler(),
-                    UserInformationEndpoint = "https://graph.facebook.com/v2.5/me?fields=id,name,email,first_name,last_name",
-                    Events = new OAuthEvents()
-                    {
-                        OnRemoteFailure = ctx => HandleRemoteLoginFailure(ctx)
-                    }
-                };
-                options.Scope.Add("email");
-
-                app.UseFacebookAuthentication(options);
-            }
-
-            if (Configuration["Authentication:MicrosoftAccount:ClientId"] != null)
-            {
-                var options = new MicrosoftAccountOptions
-                {
-                    ClientId = Configuration["Authentication:MicrosoftAccount:ClientId"],
-                    ClientSecret = Configuration["Authentication:MicrosoftAccount:ClientSecret"],
-                    Events = new OAuthEvents()
-                    {
-                        OnRemoteFailure = ctx => HandleRemoteLoginFailure(ctx)
-                    }
-                };
-
-                app.UseMicrosoftAccountAuthentication(options);
-            }
-
-            //http://www.bigbrainintelligence.com/Post/get-users-email-address-from-twitter-oauth-ap
-            if (Configuration["Authentication:Twitter:ConsumerKey"] != null)
-            {
-                var options = new TwitterOptions
-                {
-                    ConsumerKey = Configuration["Authentication:Twitter:ConsumerKey"],
-                    ConsumerSecret = Configuration["Authentication:Twitter:ConsumerSecret"]
-                    ,
-                    Events = new TwitterEvents()
-                    {
-
-                        OnRemoteFailure = ctx => HandleRemoteLoginFailure(ctx)
-                    }
-                };
-
-                app.UseTwitterAuthentication(options);
-            }
-
-            if (Configuration["Authentication:Google:ClientId"] != null)
-            {
-                var options = new GoogleOptions
-                {
-                    ClientId = Configuration["Authentication:Google:ClientId"],
-                    ClientSecret = Configuration["Authentication:Google:ClientSecret"],
-                    Events = new OAuthEvents()
-                    {
-                        OnRemoteFailure = ctx => HandleRemoteLoginFailure(ctx)
-                    }
-                };
-
-                app.UseGoogleAuthentication(options);
-            }
+            Authentication.ConfigureAuthentication(app, Configuration);
 
             //call Migrate here to force the creation of the AllReady database so Hangfire can create its schema under it
             if (!env.IsProduction())
@@ -420,15 +212,10 @@ namespace AllReady
             if (Configuration["SampleData:InsertTestUsers"] == "true")
             {
                 await sampleData.CreateAdminUser();
-            }   
+            }
         }
 
-        //Handles remote login failure typically where user doesn't give consent
-        private Task HandleRemoteLoginFailure(FailureContext ctx)
-        {
-            ctx.Response.Redirect("/Account/Login");
-            ctx.HandleResponse();
-            return Task.FromResult(0);
-        }
+
+
     }
 }
